requirements.txt
numpy==1.21.5
opencv_python==4.10.0.84
Pillow==9.4.0
PyYAML==6.0.1
torch==1.13.1
tqdm==4.64.1
